Bound first/finish rollout sample logs in the class-based and fully-async drivers
